+++ src/com/delivery/persistent/FlavourDao.java
@@ -146,7 +146,7 @@ public class FlavourDao extends Dao<Flavour> {
     }
 
     @Override
-    public int getLastSavedId() throws SQLException {
+    public long getLastSavedId() throws SQLException {
         int id = 0;
         try {
             Statement stm = mConnection.createStatement();
