+++ src/com/delivery/persistent/OrderItemOptionlRelDao.java
@@ -130,7 +130,7 @@ public class OrderItemOptionlRelDao extends Dao<OrderItemRelOptional> {
     }
 
 	@Override
-	public int getLastSavedId() throws SQLException {
+	public long getLastSavedId() throws SQLException {
 		throw new UnsupportedOperationException();
 	}
