+++ src/com/delivery/persistent/ProductDao.java
@@ -157,7 +157,7 @@ public class ProductDao extends Dao<Product> {
     }
 
     @Override
-    public int getLastSavedId() throws SQLException {
+    public long getLastSavedId() throws SQLException {
         int id = 0;
         try {
             Statement stm = mConnection.createStatement();
